requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.7
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
inflection==0.5.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.5.0
langchain==0.3.0
langchain-core==0.3.1
langchain-text-splitters==0.3.0
langsmith==0.1.121
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
packaging==24.1
protobuf==4.25.4
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.11.1
